Switch lazification approach from module splitting to function importing
So far we were splitting lazy functions out into a new module.
It was a clean approach, but we had to do most of the work at
compile-time. This adds too much overhead for code which we don't
even know whether we will use.

The new function importing changes that: we only do minimal
transformations at compile-time, then save the module to disk and
import individual functions from it at runtime. This way we can
defer much more work to runtime and keep compile-time overhead low.